Dernières corrections d'après la PR
