+++ src/app/creation/pages/creation-detail/creation-detail.component.html
@@ -5,7 +5,7 @@
 
     <div class="section-body">
       <div>
-        <img src="{{ PICTURE_URL + creationView.picture }}" alt="{{ creationView.picture }}">
+        <img src="{{ PICTURE_BASE_PATH + creationView.picture }}" alt="{{ creationView.picture }}">
       </div>
 
       <div class="creation-texts">
