+++ src/app/creation/pages/creation-detail/creation-detail.component.scss
