+++ src/app/snippet/components/snippet-list/snippet-list.component.html
@@ -1,5 +1,5 @@
 <div class="list">
 
-  <app-snippet-card class="snippet-card" (click)="sendSnippetId(snippet.id)" *ngFor="let snippet of snippets" [snippet]="snippet"></app-snippet-card>
+  <app-snippet-card class="snippet-card" (click)="sendSnippet(snippet)" *ngFor="let snippet of snippets" [snippet]="snippet"></app-snippet-card>
 
 </div>
