+++ src/app/snippet/pages/snippet/snippet.component.html
@@ -13,14 +13,10 @@
       </header>
 
       <section class="snippets-body">
-        <app-snippet-list class="snippet-list" (snippetClicked)="receivedSnippetId($event)"
+        <app-snippet-list class="snippet-list" (snippetClicked)="receiptSnippet($event)"
                         [snippets]="snippets"></app-snippet-list>
 
         <app-snippet-detail class="snippet-detail" [snippet]="activeSnippet"></app-snippet-detail>
-<!--        <router-outlet name="right-side-snip"></router-outlet>-->
-<!--        <ng-component><app-snippet-detail class="snippet-detail" [snippet]="activeSnippet"></app-snippet-detail></ng-component>-->
-<!--        <ng-component><app-snippet-list class="snippet-list" (snippetClicked)="receivedSnippetId($event)"-->
-<!--                                        [snippets]="snippets"></app-snippet-list></ng-component>-->
       </section>
     </div>
   </div>
